refactor : date 관련 유틸리티 함수 util 모듈에 정의 및 util 모듈 의존성 추가
