+++ feature/profile/create/src/commonMain/kotlin/com/whatever/caramel/feature/profile/create/ProfileCreateViewModel.kt
@@ -5,7 +5,7 @@ import com.whatever.caramel.core.domain.exception.CaramelException
 import com.whatever.caramel.core.domain.usecase.user.CreateUserProfileUseCase
 import com.whatever.caramel.core.domain.validator.UserValidator
 import com.whatever.caramel.core.domain.vo.user.Gender
-import com.whatever.caramel.core.util.createDateString
+import com.whatever.caramel.core.util.DateFormatter.createDateString
 import com.whatever.caramel.core.viewmodel.BaseViewModel
 import com.whatever.caramel.feature.profile.create.mvi.ProfileCreateIntent
 import com.whatever.caramel.feature.profile.create.mvi.ProfileCreateSideEffect
